Decouple ObjectMapper and error handling from AbstractMcpTransport
